requirements.txt
Flask
Flask-cors
Flask-SQLAlchemy
mysql-connector-python
python-dotenv
pandas
scikit-learn